UPSTREAM: mb/google/dedede: Add mainboard acpi support for GPIO PM configuration
Setting the default values for GPIO community power management, causes
issues in detecting TPM interrupts. So to avoid that GPIO PM has to be
disabled in devicetree. But for S0ix it is needed. This patch implements
a workaround in ASL code to enable GPIO PM on S0ix entry and disable it
on S0ix exit.

This patch adds the following three platform specific methods.

1. MS0X to enable power management features for GPIO communities on
low power mode entry and disables it on exit.

2. MPTS to enable power management features for GPIO communities when
preparing to sleep.

3. MWAK to disable power management features for GPIO communities on
waking up.

BUG=b:153847814
TEST=Verify S0ix is working. GPIO PM configuration is upadated on low
     power mode entry and exit.

Change-Id: I437b0badef155584191bc7d33c6febd86efccbbe
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 214c719eed83967b8f0564feca65eebb3d83f5bc
Original-Change-Id: I7225b78ab2ac5bf17f93230cd85cd21e836d807d
Original-Signed-off-by: Aamir Bohra <aamir.bohra@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41502
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2239953
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>