UPSTREAM: google/trogdor: Fix ram_code and sku_id strappings
I'm not quite sure what happened when we first added the code for
Trogdor strappings but something clearly seems to be wrong. First of
all, on newer schematics the RAM_ID_1 pin is actually pin 19, not pin
91. It only used to be 91 on rev0. Whether that was an intentional
change or someone just swapped the digits on accident at some point,
we're not quite sure anymore, but it seems to be 19 going forward so
that is what we should be programming. (ram_code wasn't used for
anything on Trogdor rev0 so we don't care about adding
backwards-compatibility for that.)

The sku_id pins are also somewhat out of whack: first of all, a new
SKU_ID_2 pin was added for rev1 that wasn't there on rev0. Second,
SKU_ID_0 is not GPIO_114. In fact, it has never been GPIO_114. I have no
idea how that number got there. Anyway, fix it. (Like with the ram_code,
SKU IDs were also not used for rev0 so we won't make this
backwards-compatible.)

BUG=none
BRANCH=none
TEST=none

Change-Id: I721102fd5b42a9721eb79b0844b6d99b7236cb47
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 238b10a5f7fb3ca6c4aca37a14f34cafa4306ddf
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia14ec74ec2f16ce2661f89d0d597a5477297ab69
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41545
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philip Chen <philipchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2211776
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>