UPSTREAM: device/resource_allocator_v3: Do not set IORESOURCE_ASSIGNED for size 0 resource
find_pci_tolm() is updated to ensure that it ignores resources that
have a zero size. This change removes the setting of resource flags to
IORESOURCE_ASSIGNED when the resource is not really allocated any
space by the allocator. It also drops the setting of base to limit
since that is not required anymore.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 468bc6cd8f765016aa4e01e20e8f020ac820cbb7
Original-Change-Id: If8c0d4bf1aa9cd6a5bdf056140f65cf2d70ed216
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41566
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Id1bf7c3d63b21393cc2fbb717c9683d73bf2ece4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2219721
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>