UPSTREAM: fit: Swap compat matching priorities for board-revX and board-skuY
Matching the same behavior change in depthcharge's FIT image code
(CL:2212466), this patch changes the order in which compat strings
involving revision and SKU numbers are matched when looking for a
compatible device tree. The most precise match (board-revX-skuY) is
still the highest priority, but after that we will now first check for
revision only (board-revX) and then for SKU only (board-skuY). The
reason for this is that SKU differentiation is often added later to a
project, so device trees for earlier revisions may not have SKU numbers
defined. So if we have a rev0 board (with sku0 as the "default SKU",
because the board only started having different SKUs with rev1) we want
it to match the board-rev0 device tree, not board-sku0 which was added
as an alias to board-rev1-sku0 to provide the best known default for
potential later revisions of that SKU.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic1951d1376ec2dcf285d51ad5586f47d20e1d71b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 11217de375598c7a6f6288d0bc04dec115e41df5
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia3cf7cbb165170e2ab0bba633fec01f9f509b874
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41633
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2226651
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>