UPSTREAM: docker/coreboot-jenkins-node: Add meson and ninja
Our jenkins instance is also used for flashrom, which can be built with
meson, a mode that we want to be able to test, so add that.

ninja can be used as a backend to both meson and cmake (which coreboot
will use to build cmocka for its unit tests) and may provide some
additional coverage. Plus it's tiny but fast.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iecda54a57f33393980d95891c23f611495ed656a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 02a1344921a6cd848fe4f766e64e602632435ea8
Original-Change-Id: If454164852303144eaa72c4071c03ee89e863318
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41731
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2219969
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>