UPSTREAM: drivers/vpd: Fix VPD speed regressions on non-x86 devices
CB:34634 expanded the VPD code to also be usable from romstage,
shuffling a few things around and adding some extra infrastructure in
the process. Unfortunately, the changes seem to have only been written
with x86 devices in mind and make coreboot always load the whole VPD
FMAP section (not just the used part) on devices where rdev_mmap() is
not a no-op.

This patch rewrites the VPD code to be based on region_device structures
that only represent the VPD area actually used (rather than the whole
FMAP section), and that only get mapped when accessed. (It would be even
better to pull this concept into the VPD decoder itself, but since that
is taken from third-party code and accesses in early stages aren't very
common, let's not go there for now.) It also moves the copying into
CBMEM to romstage so that late romstage accesses can already benefit
from it, and makes early decoding available in all stages because at
this point, why not.

Also fix a long-standing bug where the 'consumed' counter was not reset
between vpd_decode_string() calls to the RO and the RW VPD.

BUG=none
BRANCH=none
TEST=none

Change-Id: I57fab5e54caa3a16dd61a695ae483bc88089822c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5f17458cfbe90171bf6d20d9c36470ebbfb3842f
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I55a103180b290c1563e35a25496188b6a82e49ff
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41757
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2219994
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>