UPSTREAM: soc/amd/picasso: add psp_verstage
This is the main code for building coreboot's verstage as a userspace
application to run on the PSP.  It does a minimal setup of hardware,
then runs verstage_main.  It uses hardware hashing to increase the speed
and will directly reboot into recovery mode if there are any failures.

BUG=b:158124527
TEST=Build & boot trembyle

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: c7acf1666a517f0fcfec3a5f64791932ca145c45
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ia58839caa5bfbae0408702ee8d02ef482f2861c4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41816
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Change-Id: I9db79b375bfbf75dbfee77136013e536ea1299f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2292613
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>