UPSTREAM: dptf: Introduce new paradigm for configuring DPTF parameters
Currently, configuring and reviewing DPTF parameters is difficult
because DPTF tables and methods are defined in static ASL files, and are
littered with #ifdefs which both define parameters and influence
behavior (e.g., whether a method is included or not). This patch train
is an effort to bring DPTF support to our ACPI DSDT/SSDT generation
framework.

This first patch is very minimal, and includes only creation of the
DPTF device (in the DSDT).

BUG=b:143539650
TEST=compiles (later tests get more comprehensive).

Change-Id: I0ecadda5ccb00eb0b0ad7d62e632f63df3160c6d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 103bd5e4bb14fda521f64525394b2e9232297dfd
Original-Change-Id: I14df9f422c911677aeea25552ac1822a9462c58a
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41883
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2245743
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>