UPSTREAM: dptf: Add support for Fan and TSR options
DPTF has several options on how to control the fan (fine-grained speed
control, minimum speed change in percentage points, and whether or not
the DPTF device should notify the Fan if it detects low speed).
Individual TSRs can also set GTSH, which is the amount of hysteresis
inherent in the measurement, either from circuitry (if analog), or in
firmware (if digital).

BUG=b:143539650
TEST=compiles

Change-Id: Ic8b223d36044fea19bdc4faa6f84590c41a3429d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e4d8ebcef783f89fb9645d26d339dfb33065530f
Original-Change-Id: I42d789d877da28c163e394d7de5fb1ff339264eb
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41891
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2285418
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>