UPSTREAM: superio/ite: configure EC for fans to full at thermal limit
This applies to the automatic fan control mode of the environment
controller (EC). Previously the affected bit was always cleared while
the default value is 1 according to datasheets. Add a variable that can
be set per mainboard in devicetree.cb.

In the IT8783E datasheet that bit is marked as reserved.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: e693b1d549621acfd20aaad876d4d5888ad5ca99
Original-Signed-off-by: Michael Büchler <michael.buechler@posteo.net>
Original-Change-Id: Ie74102ac0d54be33558c161c9c84594d121772b1
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44165
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Change-Id: Ic0048e340cc7c19504fd787e70bd074df10dc3a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2351681
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>