UPSTREAM: ec/google/wilco: Always use current value of battery status bit
According to the Wilco EC spec the BTSC bit of PWSR is always cleared
when PWSR is read so that battery status change events are only
triggered one time. Testing of the Wilco EC has verified this behavior.
This changes the way in which the battery status change bit is used from
checking the bit state against the previous value to always issuing a
battery event when the BTSC bit is set. The other bits in PWSR indicate
state directly and do not behave like the BTSC bit.

BUG=b:157113138
TEST=Deploy on Drallion and verify that battery events are generated
BRANCH=drallion, sarien

Change-Id: Ia08bac86ae5e183049a193fe45a5f6a49fe83766
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 071182ade30da4bdba6da03ce79abf8df91ee480
Original-Signed-off-by: Mathew King <mathewk@chromium.org>
Original-Change-Id: I8fbf2ee1158ddd790b04a20b1eb27a6cce4f5c81
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42017
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2234366
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>