UPSTREAM: src/*: Update makefiles to exclude x86 code from psp-verstage
The assumption up to this point was that if the system had an x86
processor, verstage would be running on the x86 processor.  With running
verstage on the PSP, that assumption no longer holds true, so exclude
pieces of code that cause problems for verstage on the PSP.

This change will add these files to verstage only if the verstage
architecture is X86 - either 32 or 64 bit.

BUG=b:158124527
TEST=Build and boot on Trembyle

Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Commit-Id: cfc170b2133c9e6b385d218bdc5677da29275264
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I797b67394825172bd44ad1ee693a0c509289486b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/42062
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-by: Eric Peers <epeers@google.com>
Original-Reviewed-by: Rob Barnes <robbarnes@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I2d2b3740b57eb793283868bbf6138221e7584eec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2260219
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>